uradjena cela autentikacija, treba jos doterati sve i srediti auto-logout, dodati spinere, case-ove za poruke, resena autorizacija

+++ backend/routes/nekretnine.js
@@ -16,13 +16,14 @@ router.get("/api/nekretnine", (req, res, next)=>{
   });
 });
 
-router.post("/api/nekretnine", (req, res, next) => {
+router.post("/api/nekretnine", proveriAuth, (req, res, next) => {
   const nekretnina = new Nekretnina({
     naslov: req.body.naslov,
     opis: req.body.opis,
     kvadratura: req.body.kvadratura,
     cena: req.body.cena,
-    slika: req.body.slika
+    slika: req.body.slika,
+    user: req.userData.userId
   });
   console.log(nekretnina);
   nekretnina.save().then(result => {
@@ -34,10 +35,14 @@ router.post("/api/nekretnine", (req, res, next) => {
   // next();
 });
 
-router.delete("/api/nekretnine/:id", (req, res, next) => {
-  Nekretnina.deleteOne({_id: req.params.id}).then(result => {
+router.delete("/api/nekretnine/:id", proveriAuth, (req, res, next) => {
+  Nekretnina.deleteOne({_id: req.params.id, user: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Nekretnina deleted!'});
+    if(result.n > 0) {
+      res.status(200).json({message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 });
 
@@ -51,10 +56,14 @@ router.get("/api/nekretnine/:id", (req, res, next) => {
   });
 });
 
-router.put("/api/nekretnine/:id", (req, res, next) => {
-  Nekretnina.updateOne({_id: req.body.id}, req.body).then(result => {
+router.put("/api/nekretnine/:id", proveriAuth, (req, res, next) => {
+  Nekretnina.updateOne({_id: req.body.id, user: req.userData.userId}, req.body).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Update successful!'});
+    if(result.nModified > 0) {
+      res.status(200).json({message: 'Update successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 });
